+++ app/src/main/java/org/illegaller/ratabb/hishoot2i/presenter/CropActivityPresenter.java
@@ -10,6 +10,7 @@ import org.illegaller.ratabb.hishoot2i.utils.SimpleSchedulers;
 import org.illegaller.ratabb.hishoot2i.utils.UILHelper;
 import org.illegaller.ratabb.hishoot2i.utils.Utils;
 import org.illegaller.ratabb.hishoot2i.view.CropActivityView;
+import org.illegaller.ratabb.hishoot2i.view.widget.CropImageView;
 import rx.Observable;
 import rx.Subscriber;
 
@@ -38,8 +39,8 @@ public class CropActivityPresenter implements IPresenter<CropActivityView> {
     });
   }
 
-  public void performSaveCrop(Bitmap bitmap) {
-    saveObservable(bitmap).subscribe(new SimpleObserver<Uri>() {
+  public void performSaveCrop(CropImageView cropImageView) {
+    saveObservable(cropImageView).subscribe(new SimpleObserver<Uri>() {
       @Override public void onNext(Uri uri) {
         view.onResult(uri);
       }
@@ -55,18 +56,27 @@ public class CropActivityPresenter implements IPresenter<CropActivityView> {
     view.showProgress(true);
     return Observable.create(new Observable.OnSubscribe<Bitmap>() {
       @Override public void call(Subscriber<? super Bitmap> subscriber) {
-        Bitmap bitmap = UILHelper.loadImage(pathImage);
-        if (bitmap == null) subscriber.onError(new Exception("display image crop"));
-        subscriber.onNext(bitmap);
-        subscriber.onCompleted();
+        try {
+          Bitmap bitmap = UILHelper.loadImage(pathImage);
+          subscriber.onNext(bitmap);
+          subscriber.onCompleted();
+        } catch (Exception e) {
+          subscriber.onError(new Exception("display image crop"));
+        }
       }
     }).subscribeOn(schedulers.backgroundThread()).observeOn(schedulers.mainThread());
   }
 
-  Observable<Uri> saveObservable(final Bitmap bitmap) {
+  Observable<Uri> saveObservable(final CropImageView cropImageView) {
     return Observable.create(new Observable.OnSubscribe<Uri>() {
       @Override public void call(Subscriber<? super Uri> subscriber) {
         try {
+          Bitmap bitmap = null;
+          try {
+            bitmap = cropImageView.getCroppedBitmap();
+          } catch (Exception e) {
+            subscriber.onError(e);
+          }
           File file = Utils.saveTempBackgroundCrop(view.context(), bitmap);
           subscriber.onNext(Uri.fromFile(file));
           subscriber.onCompleted();