+++ app/src/main/java/org/illegaller/ratabb/hishoot2i/presenter/MainActivityPresenter.java
@@ -26,6 +26,7 @@ import org.illegaller.ratabb.hishoot2i.model.ImageReceive;
 import org.illegaller.ratabb.hishoot2i.model.template.Template;
 import org.illegaller.ratabb.hishoot2i.utils.AnimUtils;
 import org.illegaller.ratabb.hishoot2i.utils.CrashLog;
+import org.illegaller.ratabb.hishoot2i.utils.FileUtils;
 import org.illegaller.ratabb.hishoot2i.utils.ResUtils;
 import org.illegaller.ratabb.hishoot2i.utils.SimpleObserver;
 import org.illegaller.ratabb.hishoot2i.utils.UILHelper;
@@ -75,7 +76,7 @@ public class MainActivityPresenter
   }
 
   public void handleImageReceive(Uri imageUri, final String templateId) {
-    String stringUri = Utils.getStringFromUri(mView.context(), imageUri);
+    String stringUri = FileUtils.getPath(mView.context(), imageUri);
     final String sImagePath = UILHelper.stringFiles(new File(stringUri));
     templateManager.getTemplateList(TemplateManager.NO_FAV)
         .subscribe(new SimpleObserver<List<Template>>() {